+++ src/app/Http/Controllers/BaseApiController.php
@@ -1,0 +1,105 @@
+<?php
+
+namespace ByteNet\LaravelBase\app\Http\Controllers;
+
+use App\Http\Controllers\Controller;
+use Illuminate\Http\Response as IlluminateResponse;
+use Illuminate\Support\Facades\Response;
+
+class BaseApiController extends Controller
+{
+    /**
+     * @var int
+     */
+    protected $statusCode = 200;
+
+    /**
+     * Create a new controller instance.
+     */
+    public function __construct()
+    {
+        return json_encode(new \stdClass());
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getStatusCode()
+    {
+        return $this->statusCode;
+    }
+
+    /**
+     * @param mixed $statusCode
+     * @return $this
+     */
+    public function setStatusCode($statusCode)
+    {
+        $this->statusCode = $statusCode;
+
+        return $this;
+    }
+
+    /**
+     * @param string $message
+     * @return mixed
+     */
+    public function respondNotFound($message = 'Not Found')
+    {
+        return $this->setStatusCode(IlluminateResponse::HTTP_NOT_FOUND)->respondWithError($message);
+    }
+
+    /**
+     * @param string $message
+     * @return mixed
+     */
+    public function respondInternalError($message = 'Internal Error!')
+    {
+        return $this->setStatusCode(IlluminateResponse::HTTP_INTERNAL_SERVER_ERROR)->respondWithError($message);
+    }
+
+    /**
+     * @return mixed
+     */
+    public function respondCreated($message)
+    {
+        return $this->setStatusCode(IlluminateResponse::HTTP_CREATED)->respond([
+            'message' => $message,
+            'status' => 'success',
+        ]);
+    }
+
+    /**
+     * @return mixed
+     */
+    public function respondUnprocessableEntity($message)
+    {
+        return $this->setStatusCode(IlluminateResponse::HTTP_UNPROCESSABLE_ENTITY)->respondWithError($message);
+    }
+
+    /**
+     * @param $data
+     * @param array $headers
+     * @return mixed
+     */
+    public function respond($data, $headers = [])
+    {
+        return Response::json($data, $this->getStatusCode(), $headers);
+    }
+
+    /**
+     * @param $message
+     * @return mixed
+     */
+    public function respondWithError($message)
+    {
+        return $this->respond([
+            'errors' => [
+                [
+                    'messages' => [$message],
+                    'status_code' => $this->getStatusCode(),
+                ]
+            ],
+        ]);
+    }
+}